requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.11.0
    # via
    #   -r requirements.in
    #   httpx
    #   openai
asgiref==3.10.0
    # via
    #   -r requirements.in
    #   django
build==1.3.0
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2025.10.5
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
click==8.3.0
    # via
    #   -r requirements.in
    #   pip-tools
colorama==0.4.6
    # via
    #   -r requirements.in
    #   build
    #   click
    #   tqdm
distro==1.9.0
    # via
    #   -r requirements.in
    #   openai
django==5.2.7
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.16.1
    # via -r requirements.in
dotenv==0.9.9
    # via -r requirements.in
h11==0.16.0
    # via
    #   -r requirements.in
    #   httpcore
httpcore==1.0.9
    # via
    #   -r requirements.in
    #   httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   openai
    #   python-telegram-bot
idna==3.11
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
jiter==0.11.1
    # via
    #   -r requirements.in
    #   openai
openai==2.5.0
    # via -r requirements.in
packaging==25.0
    # via
    #   -r requirements.in
    #   build
pip-tools==7.5.1
    # via -r requirements.in
pydantic==2.12.3
    # via
    #   -r requirements.in
    #   openai
pydantic-core==2.41.4
    # via
    #   -r requirements.in
    #   pydantic
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   dotenv
python-telegram-bot==22.5
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
    #   openai
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   openai
typing-extensions==4.15.0
    # via
    #   -r requirements.in
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   -r requirements.in
    #   pydantic
tzdata==2025.2
    # via
    #   -r requirements.in
    #   django
wheel==0.45.1
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
